+++ besu/src/main/java/org/hyperledger/besu/cli/BesuCommand.java
@@ -1936,8 +1936,13 @@ public class BesuCommand implements DefaultCommandValues, Runnable {
         throw new ParameterException(
             commandLine, String.format("%s %s", "Checkpoint sync", errorSuffix));
       }
-      if (getDataStorageConfiguration().getDataStorageFormat().equals(DataStorageFormat.BONSAI)) {
+      if (getDataStorageConfiguration().getDataStorageFormat() == DataStorageFormat.BONSAI) {
         throw new ParameterException(commandLine, String.format("%s %s", "Bonsai", errorSuffix));
+      }
+      if (getDataStorageConfiguration().getDataStorageFormat()
+          == DataStorageFormat.X_BONSAI_ARCHIVE) {
+        throw new ParameterException(
+            commandLine, String.format("%s %s", "Bonsai archive", errorSuffix));
       }
 
       if (Boolean.TRUE.equals(privacyOptionGroup.isPrivacyMultiTenancyEnabled)