+++ besu/src/main/java/org/hyperledger/besu/controller/BesuControllerBuilder.java
@@ -86,6 +86,7 @@ import org.hyperledger.besu.ethereum.storage.keyvalue.KeyValueSegmentIdentifier;
 import org.hyperledger.besu.ethereum.trie.diffbased.bonsai.BonsaiWorldStateProvider;
 import org.hyperledger.besu.ethereum.trie.diffbased.bonsai.cache.BonsaiCachedMerkleTrieLoader;
 import org.hyperledger.besu.ethereum.trie.diffbased.bonsai.storage.BonsaiWorldStateKeyValueStorage;
+import org.hyperledger.besu.ethereum.trie.diffbased.bonsai.worldview.BonsaiArchiver;
 import org.hyperledger.besu.ethereum.trie.diffbased.common.trielog.TrieLogManager;
 import org.hyperledger.besu.ethereum.trie.diffbased.common.trielog.TrieLogPruner;
 import org.hyperledger.besu.ethereum.trie.forest.ForestWorldStateArchive;
@@ -753,6 +754,20 @@ public abstract class BesuControllerBuilder implements MiningParameterOverrides 
       trieLogManager.subscribe(trieLogPruner);
     }
 
+    // TODO - do we want a flag to turn this on and off?
+    if (DataStorageFormat.X_BONSAI_ARCHIVE.equals(
+        dataStorageConfiguration.getDataStorageFormat())) {
+      final BonsaiWorldStateKeyValueStorage worldStateKeyValueStorage =
+          worldStateStorageCoordinator.getStrategy(BonsaiWorldStateKeyValueStorage.class);
+      final BonsaiArchiver archiver =
+          createBonsaiArchiver(
+              worldStateKeyValueStorage,
+              blockchain,
+              scheduler,
+              ((BonsaiWorldStateProvider) worldStateArchive).getTrieLogManager());
+      blockchain.observeBlockAdded(archiver);
+    }
+
     final List<Closeable> closeables = new ArrayList<>();
     closeables.add(protocolContext.getWorldStateArchive());
     closeables.add(storageProvider);
@@ -817,6 +832,24 @@ public abstract class BesuControllerBuilder implements MiningParameterOverrides 
     trieLogPruner.initialize();
 
     return trieLogPruner;
+  }
+
+  private BonsaiArchiver createBonsaiArchiver(
+      final WorldStateKeyValueStorage worldStateStorage,
+      final Blockchain blockchain,
+      final EthScheduler scheduler,
+      final TrieLogManager trieLogManager) {
+    final BonsaiArchiver archiver =
+        new BonsaiArchiver(
+            (BonsaiWorldStateKeyValueStorage) worldStateStorage,
+            blockchain,
+            scheduler::executeServiceTask,
+            trieLogManager,
+            metricsSystem);
+
+    archiver.initialize();
+    LOG.info("Bonsai archiver initialised");
+    return archiver;
   }
 
   /**
@@ -1104,8 +1137,23 @@ public abstract class BesuControllerBuilder implements MiningParameterOverrides 
       final BonsaiCachedMerkleTrieLoader bonsaiCachedMerkleTrieLoader) {
     return switch (dataStorageConfiguration.getDataStorageFormat()) {
       case BONSAI -> {
+        yield new BonsaiWorldStateProvider(
+            worldStateStorageCoordinator.getStrategy(BonsaiWorldStateKeyValueStorage.class),
+            blockchain,
+            Optional.of(dataStorageConfiguration.getBonsaiMaxLayersToLoad()),
+            bonsaiCachedMerkleTrieLoader,
+            besuComponent.map(BesuComponent::getBesuPluginContext).orElse(null),
+            evmConfiguration);
+      }
+      case X_BONSAI_ARCHIVE -> {
         final BonsaiWorldStateKeyValueStorage worldStateKeyValueStorage =
             worldStateStorageCoordinator.getStrategy(BonsaiWorldStateKeyValueStorage.class);
+
+        // TODO, better integrate. Just for PoC, explicitly set our bonsai context chain head:
+        worldStateKeyValueStorage
+            .getFlatDbStrategy()
+            .updateBlockContext(blockchain.getChainHeadHeader());
+
         yield new BonsaiWorldStateProvider(
             worldStateKeyValueStorage,
             blockchain,